chore(test): organize folders
