+++ src/index.ts
@@ -1,3 +1,3 @@
 //export * from './hooks';
 export * from './library';
-export * from './utils';
+export * from './test-utils';
