+++ src/library/predicates/__tests__/predicates.test.ts
@@ -1,6 +1,6 @@
 import { Provider, bn } from 'fuels';
 import { IPayloadTransfer, IPayloadVault, Vault } from '../..';
-import { rootWallet, sendPredicateCoins, signin, newVault } from '../../../utils';
+import { rootWallet, sendPredicateCoins, signin, newVault } from '../../../test-utils';
 import { defaultConfigurable } from '../../../configurables';
 
 import { accounts } from '../../../mocks/accounts';
